+++ Runtime/Framework/IWatchable.cs
@@ -2,33 +2,33 @@ using UnityEngine.Events;
 
 namespace DT.UniStart {
   public interface IWatchable {
-    void AddListener(UnityAction f);
-    void RemoveListener(UnityAction f);
-    void AddOnceListener(UnityAction f);
-    void RemoveOnceListener(UnityAction f);
+    UnityAction AddListener(UnityAction f);
+    UnityAction RemoveListener(UnityAction f);
+    UnityAction AddOnceListener(UnityAction f);
+    UnityAction RemoveOnceListener(UnityAction f);
   }
   public interface IWatchable<T0> {
-    void AddListener(UnityAction<T0> f);
-    void RemoveListener(UnityAction<T0> f);
-    void AddOnceListener(UnityAction<T0> f);
-    void RemoveOnceListener(UnityAction<T0> f);
+    UnityAction<T0> AddListener(UnityAction<T0> f);
+    UnityAction<T0> RemoveListener(UnityAction<T0> f);
+    UnityAction<T0> AddOnceListener(UnityAction<T0> f);
+    UnityAction<T0> RemoveOnceListener(UnityAction<T0> f);
   }
   public interface IWatchable<T0, T1> {
-    void AddListener(UnityAction<T0, T1> f);
-    void RemoveListener(UnityAction<T0, T1> f);
-    void AddOnceListener(UnityAction<T0, T1> f);
-    void RemoveOnceListener(UnityAction<T0, T1> f);
+    UnityAction<T0, T1> AddListener(UnityAction<T0, T1> f);
+    UnityAction<T0, T1> RemoveListener(UnityAction<T0, T1> f);
+    UnityAction<T0, T1> AddOnceListener(UnityAction<T0, T1> f);
+    UnityAction<T0, T1> RemoveOnceListener(UnityAction<T0, T1> f);
   }
   public interface IWatchable<T0, T1, T2> {
-    void AddListener(UnityAction<T0, T1, T2> f);
-    void RemoveListener(UnityAction<T0, T1, T2> f);
-    void AddOnceListener(UnityAction<T0, T1, T2> f);
-    void RemoveOnceListener(UnityAction<T0, T1, T2> f);
+    UnityAction<T0, T1, T2> AddListener(UnityAction<T0, T1, T2> f);
+    UnityAction<T0, T1, T2> RemoveListener(UnityAction<T0, T1, T2> f);
+    UnityAction<T0, T1, T2> AddOnceListener(UnityAction<T0, T1, T2> f);
+    UnityAction<T0, T1, T2> RemoveOnceListener(UnityAction<T0, T1, T2> f);
   }
   public interface IWatchable<T0, T1, T2, T3> {
-    void AddListener(UnityAction<T0, T1, T2, T3> f);
-    void RemoveListener(UnityAction<T0, T1, T2, T3> f);
-    void AddOnceListener(UnityAction<T0, T1, T2, T3> f);
-    void RemoveOnceListener(UnityAction<T0, T1, T2, T3> f);
+    UnityAction<T0, T1, T2, T3> AddListener(UnityAction<T0, T1, T2, T3> f);
+    UnityAction<T0, T1, T2, T3> RemoveListener(UnityAction<T0, T1, T2, T3> f);
+    UnityAction<T0, T1, T2, T3> AddOnceListener(UnityAction<T0, T1, T2, T3> f);
+    UnityAction<T0, T1, T2, T3> RemoveOnceListener(UnityAction<T0, T1, T2, T3> f);
   }
 }