+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v11.0.0
 
 - **_Breaking Change_**: remove echoed events.
-  - `AddListener/RemoveListener/AddOnceListener/RemoveOnceListener` no longer return the action, and no long have `out` parameter.
+  - `AddListener/RemoveListener/AddOnceListener/RemoveOnceListener` no long have `out` parameter.
 - Feat: add `StepExecutor` family.
 - Feat: add `StateMachine` family.